PMM-12451 Required changes.
