--- agent/runner/actions/mongodb_explain_helpers.go
@@ -1,203 +1,0 @@
-// Copyright (C) 2023 Percona LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//  http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package actions
-
-import (
-	"context"
-	"strings"
-
-	"github.com/pkg/errors"
-	"go.mongodb.org/mongo-driver/bson"
-	"go.mongodb.org/mongo-driver/mongo"
-
-	"github.com/percona/pmm/version"
-)
-
-var minMongoDBVersion = version.MustParse("4.0.0-0")
-
-type explain struct {
-	Ns                 string `json:"ns"`
-	Op                 string `json:"op"`
-	Query              bson.D `json:"query,omitempty"`
-	Command            bson.D `json:"command,omitempty"`
-	OriginatingCommand bson.D `json:"originatingCommand,omitempty"`
-	UpdateObj          bson.D `json:"updateobj,omitempty"`
-}
-
-func (e explain) prepareCommand() (bson.D, error) {
-	command := e.Command
-	switch e.Op {
-	case "query":
-		if len(command) == 0 {
-			command = e.Query
-		}
-
-		if len(command) == 0 || command[0].Key != "find" {
-			var filter any
-			if len(command) != 0 && command[0].Key == "query" {
-				filter = command[0].Value
-			} else {
-				filter = command
-			}
-
-			return bson.D{
-				{Key: "find", Value: e.getCollection()},
-				{Key: "filter", Value: filter},
-			}, nil
-		}
-
-		return dropDBField(command), nil
-	case "update":
-		if len(command) == 0 {
-			command = bson.D{
-				{Key: "q", Value: e.Query},
-				{Key: "u", Value: e.UpdateObj},
-			}
-		}
-
-		return bson.D{
-			{Key: "update", Value: e.getCollection()},
-			{Key: "updates", Value: []any{command}},
-		}, nil
-	case "remove":
-		if len(command) == 0 {
-			command = bson.D{{Key: "q", Value: e.Query}}
-		}
-
-		return bson.D{
-			{Key: "delete", Value: e.getCollection()},
-			{Key: "deletes", Value: []any{command}},
-		}, nil
-	case "getmore":
-		if len(e.OriginatingCommand) == 0 {
-			return bson.D{{Key: "getmore", Value: ""}}, nil
-		}
-
-		command = e.OriginatingCommand
-
-		return dropDBField(command), nil
-	case "command":
-		command = dropDBField(command)
-
-		if len(command) == 0 || command[0].Key != "group" {
-			return command, nil
-		}
-
-		return fixReduceField(command), nil
-	// not supported
-	case "insert", "drop":
-		return nil, errors.Errorf("command %s is not supported for explain", e.Op)
-	}
-
-	return command, nil
-}
-
-func (e explain) getDB() string {
-	s := strings.SplitN(e.Ns, ".", 2)
-	if len(s) == 2 {
-		return s[0]
-	}
-
-	return ""
-}
-
-func (e explain) getCollection() string {
-	s := strings.SplitN(e.Ns, ".", 2)
-	if len(s) == 2 {
-		return s[1]
-	}
-
-	return ""
-}
-
-func dropDBField(command bson.D) bson.D {
-	for i := range command {
-		if command[i].Key != "$db" {
-			continue
-		}
-
-		if len(command)-1 == i {
-			return command[:i]
-		}
-
-		return append(command[:i], command[i+1:]...)
-	}
-
-	return command
-}
-
-func fixReduceField(command bson.D) bson.D {
-	var group bson.D
-	var ok bool
-	if group, ok = command[0].Value.(bson.D); !ok {
-		return command
-	}
-
-	for i := range group {
-		if group[i].Key == "$reduce" {
-			group[i].Value = "{}"
-			command[0].Value = group
-			break
-		}
-	}
-
-	return command
-}
-
-func explainForQuery(ctx context.Context, client *mongo.Client, query string) ([]byte, error) {
-	var e explain
-	err := bson.UnmarshalExtJSON([]byte(query), false, &e)
-	if err != nil {
-		return nil, errors.Wrapf(err, "Query: %s", query)
-	}
-
-	preparedCommand, err := e.prepareCommand()
-	if err != nil {
-		return nil, errors.Wrap(errCannotExplain, err.Error())
-	}
-	command := bson.D{{Key: "explain", Value: preparedCommand}}
-	res := client.Database(e.getDB()).RunCommand(ctx, command)
-	if res.Err() != nil {
-		return nil, errors.Wrap(errCannotExplain, res.Err().Error())
-	}
-
-	result, err := res.Raw()
-	if err != nil {
-		return nil, err
-	}
-
-	return []byte(result.String()), nil
-}
-
-func notMinMongoDBVersion(ctx context.Context, client *mongo.Client) bool {
-	var result bson.M
-	command := bson.D{{Key: "serverStatus", Value: 1}}
-	err := client.Database("check").RunCommand(ctx, command).Decode(&result)
-	if err != nil || result["version"] == nil {
-		return false
-	}
-
-	var dbVersion string
-	var ok bool
-	if dbVersion, ok = result["version"].(string); !ok {
-		return false
-	}
-	currentVersion, err := version.Parse(dbVersion)
-	if err != nil {
-		return false
-	}
-
-	return currentVersion.Less(minMongoDBVersion)
-}
